requirements.txt
streamlit
pandas
numpy
altair
Pillow
moviepy
fitparse
lxml
python-dateutil
